contador/requirements.txt
opencv-python
numpy
psycopg2-binary
